Refactor OddsScout components
Refactored the OddsScout components for improved readability and maintainability. Consider reviewing the length of individual files and refactoring further if necessary.
[skip gpt_engineer]
